+++ core/src/db/merge/MergeManagerImpl.cpp
@@ -12,6 +12,7 @@
 #include "db/merge/MergeManagerImpl.h"
 #include "db/SnapshotUtils.h"
 #include "db/SnapshotVisitor.h"
+#include "db/merge/MergeAdaptiveStrategy.h"
 #include "db/merge/MergeLayerStrategy.h"
 #include "db/merge/MergeSimpleStrategy.h"
 #include "db/merge/MergeTask.h"
@@ -20,6 +21,7 @@
 #include "utils/Log.h"
 
 #include <map>
+#include <unordered_set>
 #include <utility>
 
 namespace milvus {
@@ -73,6 +75,10 @@ MergeManagerImpl::CreateStrategy(MergeStrategyType type, MergeStrategyPtr& strat
             strategy = std::make_shared<MergeLayerStrategy>();
             break;
         }
+        case MergeStrategyType::ADAPTIVE: {
+            strategy = std::make_shared<MergeAdaptiveStrategy>();
+            break;
+        }
         default: {
             std::string msg = "Unsupported merge strategy type: " + std::to_string(static_cast<int32_t>(type));
             LOG_ENGINE_ERROR_ << msg;
@@ -91,15 +97,32 @@ MergeManagerImpl::MergeSegments(int64_t collection_id, MergeStrategyType type) {
         return status;
     }
 
+    // to avoid dead-circle, ignore failed segments
+    std::unordered_set<snapshot::ID_TYPE> failed_segments;
     while (true) {
         snapshot::ScopedSnapshotT latest_ss;
         STATUS_CHECK(snapshot::Snapshots::GetInstance().GetSnapshot(latest_ss, collection_id));
 
+        // segment must meet two conditions for merging:
+        // 1. the segment's row count is less than segment_row_count
+        // 2. the segment has no index(for any field)
         snapshot::IDS_TYPE segment_ids;
         SnapshotVisitor ss_visitor(latest_ss);
         ss_visitor.SegmentsToMerge(segment_ids);
         if (segment_ids.empty()) {
             break;  // nothing to merge
+        }
+
+        // ignore failed segments in last round
+        if (!failed_segments.empty()) {
+            for (auto iter = segment_ids.begin(); iter != segment_ids.end();) {
+                if (failed_segments.find(*iter) != failed_segments.end()) {
+                    iter = segment_ids.erase(iter);
+                } else {
+                    ++iter;
+                }
+            }
+            failed_segments.clear();
         }
 
         // collect segments info
@@ -145,7 +168,17 @@ MergeManagerImpl::MergeSegments(int64_t collection_id, MergeStrategyType type) {
         // do merge
         for (auto& segments : segment_groups) {
             MergeTask task(options_, latest_ss, segments);
-            task.Execute();
+            status = task.Execute();
+            if (!status.ok()) {
+                // merge failed, these segments will not take part in next round
+                std::string msg;
+                for (auto& id : segments) {
+                    failed_segments.insert(id);
+                    msg += std::to_string(id);
+                    msg += ",";
+                }
+                LOG_ENGINE_ERROR_ << "Failed to merge segments: " << msg << " reason: " << status.message();
+            }
         }
     }