+++ CHANGELOG.md
@@ -42,11 +42,13 @@ Please mark all changes in change log and use the issue from GitHub
 -   \#3533 Scheduler/Selector needs to judge the index type
 -   \#3621 Fix crash where getting octets information
 -   \#3626 Server crashed during search with index pq on dataset: sift-50m
+-   \#3642 Fix the bug of file not close when check sum failed.
 -   \#3652 Proto of C++ sdk is different from milvus server
 -   \#3668 Docker exit without any logs
 -   \#3672 0.11.0 docker image is 200M larger due to the un-expected installation of openblas
 -   \#3689 Client get stuck if drop collection during creating index
 -   \#3734 GET_PAGE_ENTITY sometimes failed in test_web
+-   \#3741 Inconsistent parameter names
 
 ## Feature
 -   \#2319 Redo metadata to support MVCC